Efficiency rewrite
This PR rewrites most of the implementation to reduce the overhead of
observing and scraping metrics.

The hot path for a metric observation (inc, set etc) is now
allocation-free (new label values will still result in an allocation).

During scraping, collectors now stream their metrics to a callback
(similar to how the go implemenation streams to a channel) instead of
deep-copying the whole registry.

* avoid several copies, table lookups etc of labels and metrics for each
observation (!)
* avoid deep copy of the entire metric registry, including static data
like name and help, for every scrape (!)
* deprecate reading of metric values - metrics are a one-way street
generally
* stream metrics instead of copying during scraping
* move statsd/carbon backends to separate module that must be imported
* avoid spurious thread creation when statsd/carbon is not enabled
* drop support for stdlib asyncdispatch (unmaintained)
* support for asyncdispatch can be revived in a standalone repository
by a willing maintainer
* pcre no longer needed (`-d:withoutPCRE` flag no longer needed)
* reduce imports
* reduce generated code size (fewer generics)
* print warning instead of asserting on new labels from different thread