clean up per-thread metrics
The existing implementation was confused about how running a proc in a
thread works - just because there's a pointer involved doesn't mean that
a thread-specific version of a proc is used. For the registration to
work, one would have to extract a pointer to the per-thread heap that
the nim GC uses which is not trivial.